fix: address bot review comments
- Remove unused CancellationTokenSource variable in timeout test
- Extract QueryResult, QueryColumnInfo, and MapQueryValue to shared
  QueryResultTypes.cs to eliminate duplication between QuerySqlTool
  and QueryFetchTool
